requirements.txt
# FastAPI & Server
fastapi==0.110.0
uvicorn==0.29.0
python-dev==1.0.1

# Database
SQLAlchemy==2.0.29
alembic==1.13.1

# Authentication
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# AI and Langchain
langchain==0.2.7
langchain-community==0.2.2
langchain-openai==0.1.6

# Vector Database
chromadb==0.5.0

# Utilities
pydantic==2.7.1
pydantic-settings==2.2.1







